Fixed wrong size of ColorStream buffer. More optimalizations in ScreenEmitter.

+++ core/enums/color-type-enum.h
@@ -4,7 +4,7 @@
 
 namespace Enum {
 
-	enum class ColorType : uint32_t {
+	enum class ColorType : uint8_t {
 		Red,
 		Green,
 		Blue,
